Add files via upload
Mathematically, A leap year is a year whose number is perfectly divisible by four - except for years which are both divisible by 100 and not divisible by 400. ie. The number MUST be evenly divisible by 4, but not by 100, UNLESS it is also divisible by 400.

See the first version of this program here: https://github.com/anosike1/leap-year-checker

In this version, if the user inputs nothing, the program gets the CURRENT date of the user and checks if it is a Leap Year.